Fix how tags are parsed/written

+++ mutator.go
@@ -4,7 +4,7 @@ type Mutator interface {
 	// When a field is length delimited (bytes, string, message) this method is called
 	// with the field number. If the field is a message type you can return
 	// a Mutator that will be called for each of the detected submessage fields.
-	MessageMutator(fieldNumber byte) Mutator
+	MessageMutator(fieldNumber uint64) Mutator
 
 	// Called for every field in the message.
 	//
